refacto(orm): move base orm entity and adapter in a typeorm specific folder
