+++ src/heroes/infrastructure/mock/hero.mock-adapter.ts
@@ -1,4 +1,4 @@
-import { MockAdapter } from '../../../common/infrastructure/base.mock-adapter';
+import { MockAdapter } from '../../../common/infrastructure/mock/base.mock-adapter';
 import { Hero } from '../../core/domain/hero.entity';
 import { heroEntityFactory } from '../../core/domain/hero.entity-factory';
