+++ src/heroes/infrastructure/typeorm/hero.orm-entity.ts
@@ -1,6 +1,6 @@
 import { Hero as HeroEntity } from '../../core/domain/hero.entity';
 import { Column, Entity, OneToMany } from 'typeorm';
-import { BaseOrmEntity } from '../../../common/infrastructure/base.orm-entity';
+import { BaseOrmEntity } from '../../../common/infrastructure/typeorm/base.orm-entity';
 import { Item } from '../../../items/infrastructure/typeorm/item.orm-entity';
 
 @Entity()
