+++ src/heroes/infrastructure/typeorm/heroes.adapter.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeepPartial, Repository } from 'typeorm';
-import { BaseOrmAdapter } from '../../../common/infrastructure/base.orm-adapter';
+import { BaseOrmAdapter } from '../../../common/infrastructure/typeorm/base.orm-adapter';
 import { Hero } from '../../core/domain/hero.entity';
 import { Hero as HeroOrmEntity } from './hero.orm-entity';
 import { mapHeroOrmEntityToHeroEntity } from './hero.orm-mapper';
